Testing from own laptop (windows). Changed GameScene to GameScene1 because it always starts from GameScene file. So GameScene file is the homescreen with buttons for start game, infinite game mode, and shop. Currently all the button is using temporary StartButton Image. Gameshop screen is up with a title and return button to homescreen. After a game, should be able to restart or return to homescreen.
